edited some files

+++ app/api/task/[id]/route.js
@@ -1,9 +1,14 @@
 import { NextResponse } from 'next/server';
 import connectToDB from '@/lib/connectDb';
 import { Task } from '@/models/task/task';
+import mongoose from 'mongoose';
 
 export async function GET(_, { params }) {
   try {
+    if (!mongoose.Types.ObjectId.isValid(params.id)) {
+      return NextResponse.json({ error: 'Invalid task ID' }, { status: 400 });
+    }
+
     await connectToDB();
     const task = await Task.findById(params.id).populate('assignedTo');
 
@@ -13,16 +18,25 @@ export async function GET(_, { params }) {
 
     return NextResponse.json({ data: task }, { status: 200 });
   } catch (error) {
-    return NextResponse.json({ error: 'Failed to fetch task' }, { status: 500 });
+    return NextResponse.json(
+      { error: 'Failed to fetch task' },
+      { status: 500 }
+    );
   }
 }
 
 export async function PUT(request, { params }) {
   try {
+    if (!mongoose.Types.ObjectId.isValid(params.id)) {
+      return NextResponse.json({ error: 'Invalid task ID' }, { status: 400 });
+    }
+
     const updates = await request.json();
     await connectToDB();
 
-    const updatedTask = await Task.findByIdAndUpdate(params.id, updates, { new: true });
+    const updatedTask = await Task.findByIdAndUpdate(params.id, updates, {
+      new: true,
+    }).populate('assignedTo');
 
     if (!updatedTask) {
       return NextResponse.json({ error: 'Task not found' }, { status: 404 });
@@ -33,12 +47,19 @@ export async function PUT(request, { params }) {
       { status: 200 }
     );
   } catch (error) {
-    return NextResponse.json({ error: 'Failed to update task' }, { status: 500 });
+    return NextResponse.json(
+      { error: 'Failed to update task' },
+      { status: 500 }
+    );
   }
 }
 
 export async function DELETE(_, { params }) {
   try {
+    if (!mongoose.Types.ObjectId.isValid(params.id)) {
+      return NextResponse.json({ error: 'Invalid task ID' }, { status: 400 });
+    }
+
     await connectToDB();
     const deletedTask = await Task.findByIdAndDelete(params.id);
 
@@ -51,6 +72,9 @@ export async function DELETE(_, { params }) {
       { status: 200 }
     );
   } catch (error) {
-    return NextResponse.json({ error: 'Failed to delete task' }, { status: 500 });
+    return NextResponse.json(
+      { error: 'Failed to delete task' },
+      { status: 500 }
+    );
   }
 }
